Make db indexing asynchronous
Summary:
The indexing process might take a while.
To better support long indexing times a polling endpoint was added, which returns the state of an indexing task.

In the polling endpoint we want to decrease the latency between the repo being available and the client knowing about it. For that we delay the response until a result is available or a timeout has passed. This way we maximise the chances of answering the client as soon as a database is ready, rather than the client wasting time waiting before querying the server again in a backoff strategy.

Reviewed By: simonmar

Differential Revision: D31230758

fbshipit-source-id: e8770ca63157c7c21c6c58ce2133358e9f1225ad