Proxy index request through Glass
Summary:
Proxy the request to create an incremental database through Glass to Glean.

This will simplify the LSP side of things as it will only need to talk to Glass, and will avoid requiring the LSP to know what Glean server Glass is talking to.

This endpoint only works when there is a remote connection to a Glean server as currently the `index` endpoint needs a server for indexers to talk to.

Reviewed By: zsol

Differential Revision: D31171906

fbshipit-source-id: aff90b77d1e160a4e95344bbdb49190cca2d9a34