Parallel derivation
Summary:
This is a bit of a hack but it will let us do some kinds of deriving
in parallel.

See test plan for how to run it from the command line.

Deriving will currently use 70% of the cores on the server, leaving some cores free to run the writer threads and anything else. Of course there's nothing stopping clients from easily destroying the server with multiple parallel derivations.

Reviewed By: lazamar

Differential Revision: D31272533

fbshipit-source-id: 54efdd976cedcbbb81f4557aa51022738754019f