Resume heartbeats when restarting the server
Summary:
When the server was restarted, it wasn't reinitializing the
heartbeats, and so if a client had failed a parcel while the server
was down, the server wouldn't notice and the job would hang with no
progress. This is one reason why our jobs sometimes hang for hours
after the write server is restarted.

Reviewed By: donsbot

Differential Revision: D31572867

fbshipit-source-id: 6ea71fd70625aed9063293534417ab51e2fb3597