Fix macro args as result of token pastings.
Summary:
If a macro arg is a result of a token paste, its spelling is in "scratch space". Given a minimal example like:

```
#define PASTE(x,y) x##y
#define MACRO(x) x

void foobar() {}

void f() {
  MACRO(PASTE(foo,bar))();
}
```
The call to `foobar` is missing in the Glean xrefs index because our `fixMacroLocation` function simply calls `getSpellingLoc` which retrieves the "scratch space", then the location is passed onto `ClangDB::xref` which drops the entry when it fails to map it to a real source location [here](https://fburl.com/code/utga3njl).

Reviewed By: simonmar

Differential Revision: D32015488

fbshipit-source-id: f12ff30ea58205180dc2c28052d9351bc3c54965