make C++ indexer compatible with default clang build configuration
Summary:
Clang is built by default with `-fno-rtti` as it has it's own RTTI and dynamic casting implementation. Yet we are mixing code that transitively includes folly headers using `typeid()` with code that inherits from Clang classes. Which leads to linker errors when built against no-rtti toolchain releases (including upstream), as for each class built with `-frtti` compiler emits a reference to rtti structure of the base class and that reference cannot be resolved for classes inheriting from Clang ones while linking.

The solution is conceptually simple: separate the code that includes headers containing code using `typeid()` from the code inheriting from Clang classes, and only build the former with `-frtti`. In practice it means moving isolating everything that needs to know the details of `GleanServiceAsyncClient` and `GleanClangIndexerLogger` from `index.cpp` and keeping only forward references to those in headers.

Reviewed By: simonmar

Differential Revision: D32099618

fbshipit-source-id: 8873b4f68126e7cb3a12a511ba8161d84aef5c6e