Support operators & more in DeclarationLocation
Summary:
Not all symbols in cxx have a `Name`, but `cxx.DeclarationLocation`
was trying to return a `Name` for all of them and returning no results
for some types of symbol (operators, conversion operators,
constructors, destructors, Objective C selectors).

This diff adds a new `DeclarationLocationName` with a `string` instead
of a `Name`, and fixes everything to use it.

Reviewed By: donsbot

Differential Revision: D32360324

fbshipit-source-id: 974080f3e81be77c7e0a86e754927e27a49daffe