Map predicates to their evolved counterpart
Summary:
Once we have the database content we can decide which schema will effectively evolve
which other schema. From that point on however we don't care about schemas per se
but only about predicates, so we create a map from predicate to evolved predicate.

This map will be used to decide which predicates we transform when handling queries with evolved predicates.

Reviewed By: simonmar

Differential Revision: D32353888

fbshipit-source-id: 97ed5b740c678afe32bc96a41502b04b038e23da