Check type evolution
Summary:
Add a check to assess whether predicates in a schema that is evolving another schema are in fact backward compatible.

For now we only allow new fields and new predicates to be added.

Types are treated transparently. Adding/removing/changing types only matters to the extent that this changes predicates that use those types.

Reviewed By: simonmar

Differential Revision: D32282497

fbshipit-source-id: 63a510d8aa80c8fbbe3284a538ebe1cdc798481a