Implement term transformation
Summary:
Transform a term of one type into a term of a compatible type.
If the transformation is not possible we throw an error because at this point
we should already have ruled out all illegal transformations.

If the types are identical and no transformation is needed we return `Nothing` to
avoid decoding and encoding facts that don't need to be changed.

Reviewed By: simonmar

Differential Revision: D32593203

fbshipit-source-id: 927363352218ba824041b614ee30722d29d7168d