Fix scope handling
Summary:
Fix scope handling while checking for valid evolutions.
The previous handling would fail to account for imported types.

Because type names in `TypeRef`s are fully qualified (include the schema name), it is safe to bundle all of them in one `HashMap`.
As typechecking will take care of handling out of scope types, we don't concern ourselves with scope here and consider all types to be in scope.
When creating this map we are careful to take overridden types into account.

Reviewed By: simonmar

Differential Revision: D32917697

fbshipit-source-id: aaa215a83d2b91164b3bc5cb6a57acf5036779ea