Sort derived facts by owner
Summary:
Our mapping from fact to owner is stored as an interval map, so we can
store it more compactly if there are fewer intervals. I noticed that
for some derived predicates, the order of derivation was producing a
fact mapping with lot of intervals, yet not many distinct
sets. Sorting the derived facts by owner produces a much more
efficient interval mapping.

How can we get away with sorting facts? Well, derived facts can't
currently depend on each other. When we add recursive derived facts
this will have to change, but for now it's a useful optimisation.

Reviewed By: nhawkes

Differential Revision: D32500051

fbshipit-source-id: 89f1c9155a841cf0d6a3318c98ab9fa1b078a260