Support query for evolved fact by id
Summary:
Previously we would change a `MatchFid` for a `MatchNever` whenever the type of the value was one of an old predicate evolved by a newer one.

This was done with the intent to avoid fetching facts with incorrect types.
However, that is not necessary because the client can only have retrieved the ID by querying the database for the old predicate, which will have resulted in a transformed query for the new predicate, and a transformed result shaping the fact as if it were of the old type.
Therefore the ID sent in the query is always an ID for the highest evolution of a predicate, regardless of the version required in the query.

Reviewed By: nhawkes

Differential Revision: D33098221

fbshipit-source-id: ad9c0fbc03a3189adacc175225077160e9496f54