Speed up usesOfNegation
Summary:
The function `usesOfNegation` was going through all predicates seen so far for each schema processed which is `O(n^2)`. This diff moves it to after all schemas have been typechecked, running it only once.

This improves the overall time it takes to process a complete db schema.
The target test went from 11.35s to 2.88s.

Reviewed By: simonmar

Differential Revision: D33475984

fbshipit-source-id: 7d6dc65d254864bb2c03e2684d2abba6129d2b75