Refactor of the testing framework
Summary:
Our testing framework has grown a bit gnarly over time. This diff is a start towards getting back to a sane, modular, composable, set of abstractions.

The big problem with the current state is that we have two kinds of test:

1. Tests that index some local source code, run `*.query` files and compare the results against `*.out`
2. Tests that index some local source code but then run some custom HUnit tests

The problem is that (2) only needs a subset of the functionality of (1), but the subset it needs is not cleanly separated. That's what I wanted to fix.

We need a clear notion of an "indexer" for indexing code, and then both (1) and (2) type tests can be built from these. So that's the most important new thing here: a new abstraction `Indexer`, which is similar to what we used to call `Generator`, but an `Indexer` has two differences:

* It can have its own command-line options, which are automaticall exposed via `testMain` when building a test

* Indexers can be composed (sort of). We do this in a few places, e.g. when running the Clang indexer followed by some deriving passes.

There's a lot more to do to separate things cleanly, but this is a start.

Reviewed By: lazamar

Differential Revision: D33454973

fbshipit-source-id: 248f482e16a1c326a789938b5ae81c3dd1ead60b