More housekeeping
Summary:
In the interests of having a cleaner separation between the parts of
the test framework that deal with snapshot tests (`*.query` ->
`*.out`) and the other parts of the framework, I put all the
snapshot-related functionality under `Glean.Regression.Snapshot`.

I didn't go as far as separating these into a different library,
because I realised that it actually makes sense for languages to
supply a `Driver` which can be used by both snapshot tests and other
kinds of test.

I'm writing up some docs on the framework; coming in a future diff.

Reviewed By: lazamar

Differential Revision: D33475614

fbshipit-source-id: dccb599aa1e433cfd3ef3065b996391da5516b5b