Add glean.schema_version at db creation
Summary:
When a query without predicate versions arrives, the server determines what
version these should be resolved to based on the latest version of the `all`
schema in configerator.

This means that when we create a new `all` schema all existing dbs which don't
have the newly added predicate versions would start returning empty results to
unversioned queries as it would interpret that those queries want predicates
from the latest `all` schema.

We can pin the `all` schema version to be used in a database with the
`glean.schema_version` database property. This guarantees that even if a new
`all` schema becomes available, the one to be used to resolve unversioned
queries will be the one specified in `glean.schema_version`.

This diff makes the server automatically assign the latest version of the `all`
schema to the `glean.schema_version` property at the time of db creation. If an
indexer wishes to specify a different version it can simply override it.

Reviewed By: simonmar

Differential Revision: D33528628

fbshipit-source-id: 289f31ff26758e74025d0323f795140d99390724