Make schemasCurrentVersion a Maybe
Summary:
Sometimes we don't have an 'all' schema. For example when opening a
database its stored schema will not have `all` because the thinning
process removes it.

In these cases we have relied on laziness to prevent an error being
thrown when calculating the current `all` schema version.

Using a `Maybe` here encodes explicitly the possibility of a missing
current schema version and allows us to provide a more actionable
error message in the case where it is required.

Reviewed By: simonmar

Differential Revision: D33632631

fbshipit-source-id: e3ad0045feb2d80c69ce02b8916a8b2bc6c25dba