Fix type shadowing across inherited schemas
Summary:
When a schema inherits two versions of another schema (e.g. `x.1` and `x.2`)
unversioned predicates in queries will be resolved to the predicate from the
schema with the highest version number. Unversioned types did not have that
criterion for resolution, which could cause unversioned types to be resolved to
an older-than-latest version.

Reviewed By: simonmar

Differential Revision: D33709674

fbshipit-source-id: 567e964714d931a1cc35703315bd21df7dd1c96f