Fall back when glean.schema_version is not available
Summary:
We will now ignore schema version pinning if we do not have an
`all` schema of that version.

If the client, the database, or the environment ask the server
to resolve unversioned queries against a non-existent version of
`all` that request will be ignored.

We don't throw an error here because our purpose is to continue
being able to service requests even if something weird happens
with the schema.

Reviewed By: simonmar

Differential Revision: D33951266

fbshipit-source-id: b283d9edb1790006babfdcf61241e350e0645bd5