Parallelise database validation
Summary:
Traverse the database in parallel to speed-up validation.

What takes the longest in the validation process is fetching facts from the disk.
To speed this up we will have multiple threads, each reading from a different part of the database.

The strategy here is to divide the id space in multiple sections which can be processed in parallel.
We don't divide the entire the entire id space in a few large sections because large facts may be clustered together and this division would lead to one worker taking significantly longer than all other ones.

Reviewed By: simonmar, nhawkes

Differential Revision: D33943654

fbshipit-source-id: 5537384b8ad30c075f8a38a2a392a2ee2242f6e0