Stop evolving derived predicates
Summary:
This will allow schemas to be deployed and dbs to be built for them
before they are added to `all`.

This replaces D33941002

## Justification

Imagine schema `x.2` was just created and it is not part of the `all` schema. A
schema containing only derived predicates that depends on `x` also had a new
version published, `y.2`.

Now when a query for `y.DerivedPred.1` comes in should we transform it into a
query for `y.DerivedPred.2`? Most likely this is not the intended behaviour
given `x.2` is not even part of the `all` schema yet. Maybe the database
doesn't have any facts for it at all and this evolution would break things. How
to decide? We could look at all transitive dependencies of `y.DerivedPred` to
see if any of them has facts stored in the db, but that would make the
behaviour of evolves even more subtle.

To avoid such issues we decided that a derived predicate will not be evolved.
The predicates in the query it unfolds into, however, will be.

Reviewed By: simonmar

Differential Revision: D34045187

fbshipit-source-id: 6d0ab901b74779c00eba48ebe119e22f3e788d98