Ownership introspection
Summary:
So far it hasn't been possible to directly inspect the ownership of
facts except by logging things inside the internals. This diffs adds
APIs for inspecting the ownership sets associated with facts, and for
mapping back from a UnitId to its name.

We can use this for unit tests (see `IncrementalTest`) and for
debugging purposes in the shell (future diff).

Adding the reverse mapping from UnitId to unit name will increase the
DB size but given that the name to UnitId mapping is <0.5% I don't
expect this to be a problem.

Reviewed By: lazamar

Differential Revision: D34141833

fbshipit-source-id: 75ca9a2864c509acaf94b9382ac7e4cac4b8da52