bump schemas
Summary:
This is a retake on D32450336 (https://github.com/facebookincubator/Glean/commit/28b9090ff0bff850fdbbfa5b62cca7c29bd477fc) which was reverted.  This is the first step in the migration from hack.5 to hack.6, and all.7 to all.8.

We bump all schemas depending on hack.5, and make them evolve their predecessor.

We define schema all.8 but pin client to all.7.

The update to the hack indexer is in the next commit of the stack.

1. D34169666  this commit
2. D34169665 update the indexer
3. D34349974 move clients to all.8

We could have merged 1 and 2 together (they should be deployed together and 1 isn't useful on its own, but it makes easier for review and identifying errors).

Reviewed By: nhawkes

Differential Revision: D34169666

fbshipit-source-id: 4453854292d2ee89f1de11c804e53a746b960c54