Keep list of `all` schemas available in DbSchema
Summary:
When merging the server schema with a database schema we were keeping the list of `all` schemas available in the database, which is none.

With this diff we now use the `all` schemas from the merged result.

The `schemaTypes`  and `schemaPredicates` fields expect all transitively required predicates and types to be listed there. They are used for code generation.
Given we never perform code generation with merged schemas we just make them empty here.

Reviewed By: phlalx

Differential Revision: D34581253

fbshipit-source-id: c2d5a192be3ca9b02dabe735dd45d93aae1c0760