docs: add social banner in support of Ukraine (#99)
Summary:
Our mission at Meta Open Source is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

![image](https://user-images.githubusercontent.com/12485205/156710888-e2dd4b12-a78a-4862-b79a-475f8f7d4b20.png)

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/99

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

**Static Docs Preview: glean**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D34638168/V2/glean/)|

|**Modified Pages**|

Differential Revision: D34638168

Pulled By: simonmar

fbshipit-source-id: b91f1cfe27d772d36a729aa74d32b610a5424d79