Allow more flexible dir layout for tests
Summary:
Currently tests have to be in the fixed layout

```
<root>/
  <collection>/
    predicate.query  # query to run
      <testname>/
        test.cpp   # sample source code to index
        predicate.out  # golden test output
```

but the 2-level structure is annoying for tests that have just one
test. So I'm generalising this to consider any subdirectory that
contains a ".out" file to be a test. Everything should still work, but
we can now squash out directory layout in some places.

Differential Revision: D34718822

fbshipit-source-id: 0a6414b33456964f6a6d2af1d730bd2ce8500599