Clarify the meaning of the version of `all` in Schemas
Summary:
When performing migrations we will want to commit and deploy
versions of `all` that are new but that should not be used.

We want to set the *current* version somewhere else, unrelated
to the *highest* version available.

Renaming `schemasCurrentVersion` to `schemasHighestVersion`
clarifies the fact that what we are currently using may not
be the highest version.

Reviewed By: simonmar

Differential Revision: D34752513

fbshipit-source-id: 128040a91a76e8715301fb7f33d5b11419a117e4