Support EMPTY values in unions
Summary:
Here we remove the `hs.nonempty` annotation from Thrift unions (enums coming in the next diff).

When decoding unions from bytestring we stop throwing errors if
the alternative index is out of bounds and instead we build a
value using the `EMPTY` constructor.

Note that currently there is no scenario in which an `EMPTY` value would be constructed as that would require a fact to be encoded in a way that loses some of the information it posesses. This is not the case at the moment but will be once `evolves` begins to support the addition of alternatives in sum types.

Therefore there should be no change in behaviour coming from this diff.

I added the code to handle the new `EMPTY` alternative in case expressions across the project. For value returned in the branch I chose something that seemed sensible given what was returned in the other branches.

Reviewed By: simonmar

Differential Revision: D34721746

fbshipit-source-id: 31edd669c3a377019b648a9bd242876fdddbd6f8