Make C++ indexer buildable with cabal (#96)
Summary:
This is part of https://github.com/facebookincubator/Glean/issues/95, this first patch lets us build the indexer (and the companion Haskell library), stubbing out the logger that I will quite likely need to actually implement (in addition to enabling logging in the "dumping to file" scenario, as opposed to the "talking to service" one) to figure out why the indexer doesn't seem to be producing any code fact on a trivial one-file CMake C++ project.

Quite happy to have ended up with something so small after all the adventures this took us through.

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/96

Reviewed By: nhawkes

Differential Revision: D34748365

Pulled By: simonmar

fbshipit-source-id: 5d72c853a900a7982fb16cfc2525f54449370928