Workaround #119 (clang miscompile on ubuntu 20)
We use ubuntu 22 for the clang container. However, hhvm doesn't have
images for ubuntu 22 yet, so we just won't run the hhvm tests on that
container.

To summarize:
- clang builds miscompile glean:cppexception on ubuntu 20, but passes on ubuntu 22
- hhvm runs on ubuntu 20, but not on 22 yet
Conclusion: retain the fast clang build for useful signal. For hhvm
tests , watch the regular CI until we get hhvm for ubuntu/jammy