Fix optimisation bug
Summary:
The problem here was that we start with

```
  [_1 =
     src.File.1 "fbobjc/Libraries/FBGroupMembershipReview/FBGroupsPendingMemb
ers/FBGroupPendingMembersListSectionComponent.graphql";
   _2 = buck.Target.2 {LOC_0,_,_,_};
   _3 = buck.TargetSources.3 {_2,_,_,_};
   _ = buck.Consumer.3 {_1,_3}]
```

we correctly note that `_2` should be a lookup (because it's a slow
search), but we don't identify `_3` as a lookup because we
conservatively assume that `_2` might be bound which makes it a prefix
match, and furthermore it's used in a prefix position in the
`buck.Consumer` match so doing it first might be good.

The fix is to notice that after we decide `_2` will be a lookup, we
should therefore decide that `_3` should also be a lookup because its
rhs `buck.TargetSources {_2,_,_,_}` will be a slow search.

Reviewed By: lazamar

Differential Revision: D34896366

fbshipit-source-id: 74ccbf748adcf7c6e624c3f9530856150b79a8b7