Use declaration spans not symbol spans for Hack
Summary:
The spans returned by `codemarkup.FileEntityLocations` and
`codemarkup.EntityLocation` are supposed to be whole declaration
spans, not symbol spans. Hack was doing this wrong, so this fixes it.

Thanks to the regression tests, I found that `hack.DeclarationSpan`
doesn't have facts for `ClassConstDeclaration`, so there's a temporary
workaround for that until we can fix it.

Reviewed By: nhawkes

Differential Revision: D34962247

fbshipit-source-id: 932b868a4848e95412b787cfd732def568614773