Add SymbolInfo
Summary:
SymbolInfo will be a general-purpose way for extracting
language-independent information about an Entity. At first it contains
only `kind: SymbolKind` and `isAbstract: bool`, but we can extend it
in the future to include more things.

The intention is that `SymbolInfo` extends and generalises
`SymbolKind`, so we'll depcreated the `SymbolKind` APIs in due course.

Reviewed By: nhawkes

Differential Revision: D35074963

fbshipit-source-id: ad64cf1fb9bb0f61e2f3583c1d4fc327aae1775f