Add `hack.ContainerDeclarationQName` for easier extraction of qualified names
Summary:
I'm currently prototyping a version of [WWW Facts](https://www.internalfb.com/intern/wiki/WWW_Facts/) that integrates with Glean. We would ask Glean for information about a recent public commit, and update with incremental information we got from parsing files that we learned about via Watchman.

This diff is my way of asking you if it's okay to add a convenience predicate to an existing schema. Could I just add this, or would I need to create a new schema? Would it be better if I defined the predicate as part of my client queries?

Reviewed By: simonmar

Differential Revision: D35087807

fbshipit-source-id: 774108d956807d7285d5e37d5ee2777a5464e353