Use ghc -j parallelism to speed up the Haskell build phase
There's a lot of module-level parallelism in the schema and types code.
Anything under Glean.Schema.*.Types seems to be a good candidate with
4-5x parallelism. However, we need >8G of ram to reliably use it, and at
least 4 cores.

For the CI containers on GitHub, with 2 cores/8G, there's no benefit.

HOWEVER, on the arm64 server, where we have 8 cores and 32G ram, this makes a
massive difference to the 'make' phase. E.g. with -j4 we get a 55%
reduction in build time (shaves off 1.5 hours):

https://github.com/donsbot/Glean/runs/5707150566?check_suite_focus=true
- before: 2h 58m 29s
- after: 1h 20m 43s
== 55% speedup

On my devserver, 16G ram, 6 core, we can get the make phase down to <30 mins.

```
$ time make EXTRA_GHC_OPTS="-j4 +RTS -A128m -n2m -RTS"
real	29m49.909s
user	67m55.190s
sys	2m58.837s
```

There may be a case for -j in the buck/Meta build as well, given how
much we gain in the schema compilation phase.