Use ghc -j parallelism to speed up the Haskell build phase (#127)
Summary:
There's a lot of module-level parallelism in the schema and types code.
Anything under Glean.Schema.*.Types seems to be a good candidate with
4-5x parallelism.

Cost is a bit/a lot more memory, so have to be cautious on the CI github
containers, which are 8G / 2 core. There's marginal benefit there, as the machines are memory constrained.

HOWEVER ..on the arm64 server, where we have 8 cores and 32G ram, this makes a
massive difference to the 'make' phase. E.g. with -j4:

https://github.com/donsbot/Glean/runs/5707150566?check_suite_focus=true
- before: 2h 58m 29s
- after: 1h 20m 43s
== 55% speedup of the build.

On my devserver, 16G ram, 6 core, we can get the make phase down to <30 mins.

```
$ time make EXTRA_GHC_OPTS="-j4 +RTS -A128m -n2m -RTS"
real	29m49.909s
user	67m55.190s
sys	2m58.837s
```

So this looks like it would be a good option generally for dedicated build machines.

It might suggest further work in how we generate Glean.Schema.$Lang.Types modules too.

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/127

Reviewed By: pepeiborra

Differential Revision: D35178966

Pulled By: simonmar

fbshipit-source-id: dfbc4892f48d79182fbbbd32e23dbf2c64c11978