Simplify scope handling during reorder
Summary:
We want to differentiate variables that are bound and those that
although still unbound will be part of this scope. This role was
being fulfilled by `roVisible` which was a bit confusing.
To clarify things `Scope` is now a boolean map which can tell
whether a variable is in scope and whether it is bound.

Reviewed By: simonmar

Differential Revision: D35269189

fbshipit-source-id: fb1133fdd4835f23101a83610cb37094627b4fc1