Implement if statements
Summary:
The introduction of if statements will effectively work as an optimised disjunction where one branch can be skipped.

This replaces the implementation of XOR D34686988. The rationale is that we can express XOR with if statements much more simply than representing if statements with XOR.

Patterns can be used in if statements without parenthesis, but when statements are to be used we do require brackets. (use indentation sensitivity?)

```
if src.File A
then
(   ... ;
    ... ;
)
else
(   ... ;
    ... ;
)
```

The rules for variable bindings in if statements are similar to the ones for disjunctions: if a variable is bound in the *else* branch **and** in **either** the *condition* or the *then* branch, the variable will be available outside the if statement.

Reviewed By: simonmar

Differential Revision: D35052361

fbshipit-source-id: a30af6479039eeff126887d6d27a5fa26f2a18f0