Rename evolutions to transformations
Summary:
As we will perform both forward and backward transformations
it is not right to say that a predicate is being evolved at
query time any longer, because it may in fact be transformed
into the predicate it evolves.

This diff changes names from `old` and `new` to `requested` and
`available` or `from` and `to` as this is more correct and identifies
the parties more clearly

This diff renames the query transformation process from evolution to transformation.
The transformations themselves arise from schema evolutions.

This diff does not change any logic

Reviewed By: simonmar

Differential Revision: D35316703

fbshipit-source-id: 53be195b5f9095fc989541915178da099698c32d