Don't index spans into a non-physical file.
Summary:
Some built-in function declarations such as global `operator new` as well as compiler instrinsic typedef declarations such as `__int128_t` show up in the Clang AST. These declarations don't have a corresponding physical file, and we index the file as `"<builtin>"`. The source range we construct for this is:
```
{
  file: "<builtin>",
  lineBegin: 1,
  columnBegin: 1,
  lineEnd: 1,
  columnEnd: 1
}
```
This isn't ideal, but it works okay today. Glass doesn't perform any processing on `Range`s and simply propagates them to clients.

With the addition of the name span, the resulting span is `{ start: 0, length: 0 }`. This is more problematic than a `Range` because Glass tries to compute the `Range` from the `ByteSpan` using `src.FileLines` of the corresponding file which is `"<builtin>"`! In this case, Glass simply returns a default value of `Range`, with all `0`s. Since Glean/Glass is supposed to yield 1-based line/column numbers, all `0`s` is an invalid `Range` that VSCode does not know how to handle.

For now, I propose that we simply not index name spans that don't have a corresponding physical file.

The net effect of this for now is that we lose xrefs to these odd builtin declarations. Practically, this doesn't affect the code navigation experience since we can't follow these xrefs to a builtin file anyway. From an idealistic, correctness perspective, we could make a new version of `DeclarationLocationNameSpan` with `maybe` fields for `file` and `span` depending on whether a `DeclarationNameSpan` entry exists. I don't feel this is worth the effort at least for now, and would be more inclined to considering it during the next major schema migration.

Reviewed By: simonmar

Differential Revision: D35421457

fbshipit-source-id: e8034fcc8ab3273041cab60f79db4b7770a0b9ec