Make Glean.Test.Regression.Indexer work with an arbitrary Backend
Summary: This generalisation was necessary so that we can have indexing work with `glean index --service glean.write` which would involve pointing the indexer at the given service rather than localhost. It might not be all that useful but it wasn't hard to make it work and it seems better than restricting this use case to `--db-root` only.

Reviewed By: pepeiborra

Differential Revision: D35612705

fbshipit-source-id: 14bb683e83bdd781c048cce8b26b20870f7ba76b