Fix --replace
Summary:
I inadvertently broke `--replace` in D35578719 because the source
files are now under `buck-out` so it doesn't update the copies in the
repository.

The quickest fix is just to add an argument to `--replace` to specify
the location of the golden files. This is not ideal but we don't do
`--replace` all that often and the other options I could think of were
all worse. Better ideas welcome.

Reviewed By: pepeiborra

Differential Revision: D35780405

fbshipit-source-id: 3e6e3a3cb318ea41b3dd35bc57e24aaac3efb3d7