Don't include evolves dependencies in generated schema code
Summary:
Removes 30k lines of unnecessary generated code, and should save a lot
of compile time.

When generating the code from the schema, we want to include all the
predicates that are inherited by the "all" schema. But this
calculation was done using a graph that also contained edges
corresponding to `evolves` directives, so the resulting code contained
a lot more predicates than we actually wanted.

Reviewed By: pepeiborra

Differential Revision: D35842336

fbshipit-source-id: 8f2433b2aede521e7b9170dfc09c5c7d860797e0