Fix broken test
Summary:
The test was failing in opt mode because the binary version in line 4 includes the revision.
The test broke in D35778829 (https://github.com/facebookincubator/Glean/commit/4392f36c0274139f71d05efeef6fe9461b56364e) when the expected output was regenerated in dev mode.

To fix, restore the regular expression that was used in the expected output to work around this difference between dev and opt mode (note that TestX runs the test in both modes, but Phabricator only exercises dev mode)

Reviewed By: phlalx

Differential Revision: D35928227

fbshipit-source-id: a28eaa93774e31329234ce14f52e9586d586a168