Per-language codemarkup.EntityKind predicates
Summary:
In T117030155 we see that the partial glean dbs used for sketching
lacked some of the information required to compute `isAbstract` in
`codemarkup.EntityInfo`.  Since we only need `EntityKind`, compute
that directly, and then use that to guide computation of
`isAbstract`.  For many languages `isAbstract` is unconditionally true
and we don't really need LangEntityKind at all; the exceptions right
now seem to be C++ and Hack.

Reviewed By: simonmar

Differential Revision: D35795030

fbshipit-source-id: 38f3b7e36d7f08c0829cc237af9cbdcfa65c1b4d