Avoid the minimum 1s wait per derivation
Summary:
We currently wait at least 1s for each derivation to finish, due to
the async nature of the API. This affects tests quite badly, which
typically have to do 5-10 derivations which would otherwise run very
fast.

The fix is to make the API partially synchronous: we wait up to 1s for
a new derivation to finish, before falling back to async polling.

Reviewed By: pepeiborra, lazamar

Differential Revision: D35931779

fbshipit-source-id: 1ca7cbd76817a11a5993eacbfd8fff36f59f32a8