glean-clang: support deriving in addition to indexing
This patch also automates dumping the inventory and loading the indexed data,
using the glean library, and this can all be performed against a local or remote
glean database.

Using the same example C++ project as in 7945ad2, I tested this patch by doing:

``` sh
$ mkdir /gleandb
$ cabal run glean -- \
        --db-root /gleandb \
        --db-schema dir:/Glean/glean/schema/source \
        index --repo foo/0 cpp <path to C++ project directory>
```

and then inspecting the resulting database in `glean shell`, making sure the
"raw" facts as well as the derived ones were there. This requires having the
`clang-index` and `clang-derive` executables in PATH, or alternatively to
specify the paths explicitly (see `glean index cpp --help`).