It's time to stop supporting DBVersion 2
Summary:
In version 3 we switched from writing fact IDs as fixed int64s to
variable-width nats, so that we could more easily support iteration in
order of fact ID.

We haven't created version 2 DBs for over a year now, so let's remove
support for it.

Removing support for writing version 2 DBs will instantly kill a bunch
of tests, include 1/3 of the tests run by `AngleTest.hs`, so it will
improve our CI time to signal.

Reviewed By: nhawkes

Differential Revision: D36094480

fbshipit-source-id: 025cdafd98b9f0a987966230751000ad48c89526