Fix validate
Summary:
There were a number of bugs in the parallel validator.

1. Exceptions were not being propagated from the workers to the main
thread, so certain kinds of validation failure were being silently
ignored, making it appear as if validation had succeeded. Fixed by
using `get()` on the future instead of `wait()`.

2. `fail()` didn't throw an exception, so it would continue after
failing, possibly leading to an actual exception (which would be
ignored), or a crash. We want to throw exceptions instead, but we also
want all the worker threads to stop when one of them has found a
failure, so there's a wrapper class for `fail()` now.

3. The size of the `locks` array was one element too short in many
cases, leading to an out of bounds exception (which was silently
ignored previously).

2. The way `limit` was implemented was a bit wrong, in that we would
validate a random set of `limit` facts. I've made it so that we always
validate the first `limit` facts of the DB.

Reviewed By: malanka

Differential Revision: D36340696

fbshipit-source-id: 3a15c0f4f57c5882e9e0f5876ddcc870644530e1