Handle empty spans for conversions from `Src.ByteSpan` to `Glass.Range`.
Summary:
The current implementation of the conversion from `Src.ByteSpan` to `Glass.Range` goes through utilities that deal with `Src.Range`. This results in correct answers for the most part, except for empty spans. This is due to the fact that a `Src.Range`, which has inclusive begin/end, cannot represent empty spans at all.

Since `Src.ByteSpan` and `Glass.Range` are both capable to representing empty spans, I propose to perform the transformation directly without going through `Src.Range` so that the answer is correct even for empty spans.

Reviewed By: simonmar

Differential Revision: D36579840

fbshipit-source-id: dda7f60f1a19d2ad47c1727f8785e48219e11e3b