external indexer: respect indexerOutput parameter in Json case
Otherwise the JSON files are written to a temporary directory that is _always_
deleted once the data has been loaded. Instead, this patches creates a dedicated
directory under whatever path is specified for `indexerOutput`, and it is then
up to the callers of (JSON-based) external indexers to decide whether all the
temporary files generated under `indexerOutput` should be kept or deleted.