Teach gen-schema to show predicate dependency graphs
Summary:
In a few situations it can be useful to know the direct and transitive dependencies of a predicate.

If a derived predicate suddenly has a lot fewer facts than expected we can filter the `glean_database_stats` scuba table by the predicate's dependencies to try and find which predicate is the source of the shrinking.

We can see a list of transitive dependencies by using the `--topsort` flag and not specifying a `--depth` and find out what the predicate is used for with `--reverse-deps`.

pepeiborra was also interested in perhaps using it to automate derivation triggering in Skycastle.

afterallwhynot

Reviewed By: simonmar

Differential Revision: D36641669

fbshipit-source-id: c26d7fa78d4a576d5c2fac9e870641e11ad855b4