Cross-repo symbol ids
Summary: At the moment the code assumes that symbols in the e.g. www.hack db refer to symbols in the www repo. This isn't true fro cross-repo reference. This makes symbol id requires a (repo, path) to the symbol to ensure callers check where the symbol is located, and pass *that* repo, not the repo of the database

Reviewed By: simonmar

Differential Revision: D36416448

fbshipit-source-id: e05d7a424a3711caa3d06f26a9c5aeb678a850cb