external indexer: respect indexerOutput parameter in Json case (#194)
Summary:
Otherwise the JSON files are written to a temporary directory that is _always_
deleted once the data has been loaded. Instead, this patches creates a dedicated
directory under whatever path is specified for `indexerOutput`, and it is then
up to the callers of (JSON-based) external indexers to decide whether all the
temporary files generated under `indexerOutput` should be kept or deleted.

(AFAIK, neither the `:index` command in `glean shell` nor the `glean index` CLI command support keeping the temp files around for now, but this will be trivial to implement if all indexers follow the convention to stick all the generated files under `indexerOutput`.)

cc simonmar

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/194

Reviewed By: pepeiborra

Differential Revision: D36773637

Pulled By: simonmar

fbshipit-source-id: 8f59d3e2dd5e33dff163b7e6317f4fdf89f37726