Index hie files directly
Summary:
This removes the need for the caller to create an intermediate hiedb - now the indexer can be pointed to a file tree containing `.hie` files directly and it will create the hiedb internally. This reduces complexity both in the internal and open source use cases.

Note that the Sigma indexer still uses an intermediate hiedb because the Sigma buck rules are designed to cache the hiedb, which is shared with the LSP server.

Reviewed By: simonmar

Differential Revision: D36632635

fbshipit-source-id: 2111267881b6293035f93c1ac5de942a7f765c77